+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ by Matt Tropiano et al. (see AUTHORS.txt)
 Changed in 0.1.1
 ----------------
 
+- `Added` Some event polling methods, changed main loop to wait for events with timeout.
 - `Changed` Removed some calls to init() that would make no sense.
 - `Changed` Renamed the context setting.